requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
pytest-django
isort
flake8
flake8-quotes